break up pg_dml into pg_dml, pg_dml_read, and pg_dml_write and address additional PR comments
