+++ src/k2/connector/yb/pggate/pg_dml.h
@@ -182,7 +182,9 @@ class PgDml : public PgStatement {
   // - "target_desc_" is the table descriptor where data will be read from.
   // - "targets_" are either selected or returned expressions by DML statements.
   PgTableDesc::ScopedRefPtr target_desc_;
-  std::vector<PgExpr*> targets_;
+  
+  // use unique_ptr here to make sure that we release the memory for PgExprs when we are done with the statement
+  std::vector<std::unique_ptr<PgExpr>> targets_;
 
   // bind_desc_ is the descriptor of the table whose key columns' values will be specified by the
   // the DML statement being executed.
@@ -247,149 +249,6 @@ class PgDml : public PgStatement {
   //
   // These members are not used internally by the statement and are simply a utility for computing
   // the tuple id (ybctid).
-};
-
-//--------------------------------------------------------------------------------------------------
-// DML_READ
-//--------------------------------------------------------------------------------------------------
-// Scan Scenarios:
-//
-// 1. SequentialScan or PrimaryIndexScan (class PgSelect)
-//    - We does not have a separate table for PrimaryIndex.
-//    - The target table descriptor, where data is read and returned, is the main table.
-//    - The binding table descriptor, whose column is bound to values, is also the main table.
-//
-// 2. IndexOnlyScan (Class PgSelectIndex)
-//    - This special case is optimized where data is read from index table.
-//    - The target table descriptor, where data is read and returned, is the index table.
-//    - The binding table descriptor, whose column is bound to values, is also the index table.
-//
-// 3. IndexScan SysTable / UserTable (Class PgSelect and Nested PgSelectIndex)
-//    - We will use the binds to query base-ybctid in the index table, which is then used
-//      to query data from the main table.
-//    - The target table descriptor, where data is read and returned, is the main table.
-//    - The binding table descriptor, whose column is bound to values, is the index table.
-
-class PgDmlRead : public PgDml {
- public:
-  // Public types.
-  typedef scoped_refptr<PgDmlRead> ScopedRefPtr;
-  typedef std::shared_ptr<PgDmlRead> SharedPtr;
-
-  // Constructors.
-  PgDmlRead(PgSession::ScopedRefPtr pg_session, const PgObjectId& table_id,
-           const PgObjectId& index_id, const PgPrepareParameters *prepare_params);
-  virtual ~PgDmlRead();
-
-  StmtOp stmt_op() const override { return StmtOp::STMT_SELECT; }
-
-  virtual CHECKED_STATUS Prepare() = 0;
-
-  // Allocate binds.
-  virtual void PrepareBinds();
-
-  // Set forward (or backward) scan.
-  void SetForwardScan(const bool is_forward_scan);
-
-  // Bind a column with an EQUALS condition.
-  CHECKED_STATUS BindColumnCondEq(int attnum, PgExpr *attr_value);
-
-  // Bind a range column with a BETWEEN condition.
-  CHECKED_STATUS BindColumnCondBetween(int attr_num, PgExpr *attr_value, PgExpr *attr_value_end);
-
-  // Bind a column with an IN condition.
-  CHECKED_STATUS BindColumnCondIn(int attnum, int n_attr_values, PgExpr **attr_values);
-
-  // Execute.
-  virtual CHECKED_STATUS Exec(const PgExecParameters *exec_params);
-
-  void SetCatalogCacheVersion(const uint64_t catalog_cache_version) override {
-    DCHECK_NOTNULL(read_req_)->catalog_version = catalog_cache_version;
-  }
-
-  protected:
-   // Allocate column variable.
-  std::shared_ptr<SqlOpExpr> AllocColumnBindVar(PgColumn *col) override;
-  std::shared_ptr<SqlOpCondition> AllocColumnBindConditionExprVar(PgColumn *col);
-
-  // Allocate variable for target.
-  std::shared_ptr<SqlOpExpr> AllocTargetVar() override;
-
-  // Allocate column expression.
-  std::shared_ptr<SqlOpExpr> AllocColumnAssignVar(PgColumn *col) override;
-
-  // Delete allocated target for columns that have no bind-values.
-  CHECKED_STATUS DeleteEmptyPrimaryBinds();
-
-  // References read request from template operation.
-  std::shared_ptr<SqlOpReadRequest> read_req_ = nullptr;
-};
-
-//--------------------------------------------------------------------------------------------------
-// DML WRITE - Insert, Update, Delete.
-//--------------------------------------------------------------------------------------------------
-
-class PgDmlWrite : public PgDml {
- public:
-  // Abstract class without constructors.
-  virtual ~PgDmlWrite();
-
-  // Prepare write operations.
-  virtual CHECKED_STATUS Prepare();
-
-  // Setup internal structures for binding values during prepare.
-  void PrepareColumns();
-
-  // force_non_bufferable flag indicates this operation should not be buffered.
-  CHECKED_STATUS Exec(bool force_non_bufferable = false);
-
-  void SetIsSystemCatalogChange() {
-      ysql_catalog_change_ = true;
-  }
-
-  void SetCatalogCacheVersion(const uint64_t catalog_cache_version) override {
-    ysql_catalog_version_ = catalog_cache_version;
-  }
-
-  int32_t GetRowsAffectedCount() {
-    return rows_affected_count_;
-  }
-
-  CHECKED_STATUS SetWriteTime(const uint64_t write_time);
-
- protected:
-  // Constructor.
-  PgDmlWrite(PgSession::ScopedRefPtr pg_session,
-             const PgObjectId& table_id,
-             bool is_single_row_txn = false);
- 
-  // Allocate write request.
-  void AllocWriteRequest();
-
-  // Allocate column expression.
-  std::shared_ptr<SqlOpExpr> AllocColumnBindVar(PgColumn *col) override;
-
-  // Allocate target for selected or returned expressions.
-  std::shared_ptr<SqlOpExpr> AllocTargetVar() override;
-
-  // Allocate column expression.
-  std::shared_ptr<SqlOpExpr> AllocColumnAssignVar(PgColumn *col) override;
-
-  // Delete allocated target for columns that have no bind-values.
-  CHECKED_STATUS DeleteEmptyPrimaryBinds();
-
-  std::shared_ptr<SqlOpWriteRequest> write_req_ = nullptr;
-
-  bool is_single_row_txn_ = false; // default.
-
-  int32_t rows_affected_count_ = 0;
-
-  bool ysql_catalog_change_ = false;
-
-  uint64_t ysql_catalog_version_ = 0;
-
-  private:
-  virtual std::unique_ptr<PgWriteOpTemplate> AllocWriteOperation() const = 0;
 };
 
 }  // namespace gate
