+++ src/k2/connector/yb/pggate/pg_dml_write.h
@@ -46,32 +46,82 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 //
 
-#ifndef CHOGORI_GATE_DML_UPDATE_H
-#define CHOGORI_GATE_DML_UPDATE_H
+#ifndef CHOGORI_GATE_PG_DML_WRITE_H
+#define CHOGORI_GATE_PG_DML_WRITE_H
 
-#include "yb/pggate/pg_dml_write.h"
+#include "yb/pggate/pg_dml.h"
 
 namespace k2pg {
 namespace gate {
 
-class PgUpdate : public PgDmlWrite {
+//--------------------------------------------------------------------------------------------------
+// DML WRITE - Insert, Update, Delete.
+//--------------------------------------------------------------------------------------------------
+
+class PgDmlWrite : public PgDml {
  public:
-  // Public types.
-  typedef scoped_refptr<PgUpdate> ScopedRefPtr;
+  // Abstract class without constructors.
+  virtual ~PgDmlWrite();
 
-  // Constructors.
-  PgUpdate(PgSession::ScopedRefPtr pg_session, const PgObjectId& table_id, bool is_single_row_txn)
-      : PgDmlWrite(std::move(pg_session), table_id, is_single_row_txn) {}
+  // Prepare write operations.
+  virtual CHECKED_STATUS Prepare();
 
-  StmtOp stmt_op() const override { return StmtOp::STMT_UPDATE; }
+  // Setup internal structures for binding values during prepare.
+  void PrepareColumns();
+
+  // force_non_bufferable flag indicates this operation should not be buffered.
+  CHECKED_STATUS Exec(bool force_non_bufferable = false);
+
+  void SetIsSystemCatalogChange() {
+      ysql_catalog_change_ = true;
+  }
+
+  void SetCatalogCacheVersion(const uint64_t catalog_cache_version) override {
+    ysql_catalog_version_ = catalog_cache_version;
+  }
+
+  int32_t GetRowsAffectedCount() {
+    return rows_affected_count_;
+  }
+
+  CHECKED_STATUS SetWriteTime(const uint64_t write_time);
+
+ protected:
+  // Constructor.
+  PgDmlWrite(PgSession::ScopedRefPtr pg_session,
+             const PgObjectId& table_id,
+             bool is_single_row_txn = false);
+ 
+  // Allocate write request.
+  void AllocWriteRequest();
+
+  // Allocate column expression.
+  std::shared_ptr<SqlOpExpr> AllocColumnBindVar(PgColumn *col) override;
+
+  // Allocate target for selected or returned expressions.
+  std::shared_ptr<SqlOpExpr> AllocTargetVar() override;
+
+  // Allocate column expression.
+  std::shared_ptr<SqlOpExpr> AllocColumnAssignVar(PgColumn *col) override;
+
+  // Delete allocated target for columns that have no bind-values.
+  CHECKED_STATUS DeleteEmptyPrimaryBinds();
+
+  std::shared_ptr<SqlOpWriteRequest> write_req_ = nullptr;
+
+  bool is_single_row_txn_ = false; // default.
+
+  int32_t rows_affected_count_ = 0;
+
+  bool ysql_catalog_change_ = false;
+
+  uint64_t ysql_catalog_version_ = 0;
 
   private:
-  std::unique_ptr<PgWriteOpTemplate> AllocWriteOperation() const override {
-    return target_desc_->NewPgsqlUpdate(client_id_, stmt_id_);
-  } 
+  virtual std::unique_ptr<PgWriteOpTemplate> AllocWriteOperation() const = 0;
 };
 
 }  // namespace gate
 }  // namespace k2pg
 
-#endif //CHOGORI_GATE_DML_UPDATE_H    
+#endif //CHOGORI_GATE_PG_DML_WRITE_H
